requirements.txt
langchain
langgraph
groq
arxiv
requests
streamlit
PyPDF2
python-dotenv
tiktoken
langchain_groq
beautifulsoup4
matplotlib
wordcloud